+++ hybrid_pool_executor/base.py
@@ -1,21 +1,83 @@
 from abc import ABC, abstractmethod, abstractclassmethod
 from concurrent.futures._base import Executor
-from concurrent.futures._base import Future as _Future
+from concurrent.futures._base import Future
 from dataclasses import dataclass, field
 from queue import SimpleQueue
-from typing import Any, Callable, Dict, Hashable, Optional, Tuple
+from typing import Any, Callable, Coroutine, Dict, Optional, Tuple, Type, Union
+from hybrid_pool_executor.typing import ActionFlag, ACT_NONE, Function
 
-from hybrid_pool_executor.typing import (
-    ActionFlag,
-    ACT_NONE,
-    WorkerMode,
-    WORKER_MODE_THREAD,
-)
+
+class CancelledError(Exception):
+    pass
+
 
 """
 For python 3.7+, there is no significant speed/size difference between object,
 dataclass and namedtuple.
 """
+
+
+@dataclass
+class BaseAction(ABC):
+    """The base dataclass of action.
+
+    Actions are objects used to transfer information between worker(s) and manager(s)
+    through request/response queue.
+
+    BaseAction is regarded as a abstract class and should not be initialized directly.
+    """
+
+    flag: ActionFlag = ACT_NONE
+    message: Optional[str] = None
+    task_id: Optional[str] = None
+    worker_id: Optional[str] = None
+    result: Any = None
+    exception: Optional[BaseException] = None
+
+    def add_flag(self, flag: ActionFlag):
+        self.flag |= flag
+
+    def match(self, *flags) -> bool:
+        if len(flags) == 1 and isinstance(flags[0], (tuple, list, set)):
+            flags = flags[0]
+        for flag in flags:
+            if self.flag & flag:
+                return True
+        return False
+
+
+class BaseAsyncFutureInterface(ABC):
+    @abstractclassmethod
+    async def cancel(cls, *args, **kwargs) -> bool:
+        pass
+
+    @abstractclassmethod
+    async def result(cls, *args, **kwargs) -> bool:
+        pass
+
+    @abstractclassmethod
+    async def set_result(cls, *args, **kwargs):
+        pass
+
+    @abstractclassmethod
+    async def set_exception(cls, *args, **kwargs):
+        pass
+
+
+@dataclass
+class BaseTask(ABC):
+    """The base dataclass of task.
+
+    Calls are objects used to carry functions to worker(s).
+
+    BaseCall is regarded as a abstract class and should not be initialized directly.
+    """
+
+    name: str
+    func: Function
+    args: Tuple[Any, ...] = ()
+    kwargs: Dict[str, Any] = field(default_factory=dict)
+    cancelled: bool = False
 
 
 @dataclass
@@ -25,7 +87,7 @@ class BaseWorkerSpec(ABC):
     BaseWorkerSpec is regarded as a abstract class and should not be initialized directly.
 
     :param name: Name of worker.
-    :type name: Hashable
+    :type name: str
 
     :param work_queue: The queue for sending task item.
     :type work_queue: SimpleQueue
@@ -61,119 +123,15 @@ class BaseWorkerSpec(ABC):
     :type max_cons_err_count: int, optional
     """
 
-    name: Hashable
+    name: str
     task_bus: SimpleQueue
     request_bus: SimpleQueue
     response_bus: SimpleQueue
-    daemon: bool = True
     idle_timeout: float = 60.0
     wait_interval: float = 0.1
     max_task_count: int = 12
     max_err_count: int = 3
     max_cons_err_count: int = -1
-
-
-@dataclass
-class BaseAction(ABC):
-    """The base dataclass of action.
-
-    Actions are objects used to transfer information between worker(s) and manager(s)
-    through request/response queue.
-
-    BaseAction is regarded as a abstract class and should not be initialized directly.
-    """
-
-    flag: ActionFlag = ACT_NONE
-    message: Optional[str] = None
-    task_id: Optional[Hashable] = None
-    worker_id: Optional[Hashable] = None
-    worker_mode: WorkerMode = WORKER_MODE_THREAD
-    result: Any = None
-    exception: Optional[BaseException] = None
-
-    def add_flag(self, flag: ActionFlag):
-        self.flag |= flag
-
-    def match(self, *flags) -> bool:
-        if len(flags) == 1 and isinstance(flags[0], (tuple, list, set)):
-            flags = flags[0]
-        for flag in flags:
-            if self.flag & flag:
-                return True
-        return False
-
-
-@dataclass
-class BaseCall(ABC):
-    """The base dataclass of call item.
-
-    Calls are objects used to carry functions to worker(s).
-
-    BaseCall is regarded as a abstract class and should not be initialized directly.
-    """
-
-    name: Hashable
-    func: Callable
-    args: Tuple[Any, ...] = ()
-    kwargs: Dict[str, Any] = field(default_factory=dict)
-    work_mode: WorkerMode = WORKER_MODE_THREAD
-    cancelled: bool = False
-
-
-class BaseAsyncFutureInterface(ABC):
-    @abstractclassmethod
-    async def cancel(cls, *args, **kwargs) -> bool:
-        pass
-
-    @abstractclassmethod
-    async def result(cls, *args, **kwargs) -> bool:
-        pass
-
-    @abstractclassmethod
-    async def set_result(cls, *args, **kwargs):
-        pass
-
-    @abstractclassmethod
-    async def set_exception(cls, *args, **kwargs):
-        pass
-
-
-class BaseFuture(_Future, ABC):
-    @abstractmethod
-    def cancel(self) -> bool:
-        pass
-
-    @abstractmethod
-    def cancelled(self) -> bool:
-        pass
-
-    @abstractmethod
-    def running(self) -> bool:
-        pass
-
-    @abstractmethod
-    def done(self) -> bool:
-        pass
-
-    @abstractmethod
-    def add_done_callback(self, fn: Callable[["BaseFuture"], Any]):
-        pass
-
-    @abstractmethod
-    def result(self, timeout: Optional[float] = None):
-        pass
-
-    @abstractmethod
-    def set_running_or_notify_cancel(self):
-        pass
-
-    @abstractmethod
-    def set_result(self, result: Any):
-        pass
-
-    @abstractmethod
-    def set_exception(self, exception: Exception):
-        pass
 
 
 class BaseWorker(ABC):
@@ -194,4 +152,33 @@ class BaseWorker(ABC):
         pass
 
 
+@dataclass
+class BaseManagerSpec(ABC):
+    mode: str
+    num_workers: int = -1
+    incremental: bool = True
+    name_pattern: str = "Manager-{manager}"
+    worker_name_pattern: str = "Worker-{worker}"
+    worker_daemon: bool = True
+    worker_idle_timeout: float = 60.0
+    worker_wait_interval: float = 0.1
+    worker_max_task_count: int = 12
+    worker_max_err_count: int = 3
+    worker_max_cons_err_count: int = -1
+
+
+class BaseManager(ABC):
+    pass
+
+
+BaseFuture = Future
 BaseExecutor = Executor
+
+
+@dataclass
+class ModuleSpec:
+    name: str
+    manager_class: Type[BaseManager]
+    manager_spec_class: Type[BaseManagerSpec]
+    worker_class: Type[BaseWorker]
+    worker_spec_class: Type[BaseWorkerSpec]
